dom/tests/mochitest/bugs/test_bug862540.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=862540
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 862540</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Bug 862540 **/
   window.status = 5;
   is(window.status, "5", "Should be able to get 5");

   window.status = { toString: function() { return "foo"; } };
   ok(window.status === "foo", "Should be coercing window.status to string");

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=862540">Mozilla Bug 862540</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
